Split AI chat remote wire decoding
